rename SizeChaneListener to SizeChangedListener
