+++ yandex-mapkit-kmp/src/iosMain/kotlin/ru/sulgik/mapkit/map/MapWindow.kt
@@ -24,11 +24,11 @@ public actual class MapWindow internal constructor(private val nativeMapWindow: 
     public actual val map: Map
         get() = nativeMapWindow.map.toCommon()
 
-    public actual fun addSizeChangeListener(listener: SizeChangeListener) {
+    public actual fun addSizeChangeListener(listener: SizeChangedListener) {
         nativeMapWindow.addSizeChangedListenerWithSizeChangedListener(listener.toNative())
     }
 
-    public actual fun removeSizeChangeListener(listener: SizeChangeListener) {
+    public actual fun removeSizeChangeListener(listener: SizeChangedListener) {
         nativeMapWindow.removeSizeChangedListenerWithSizeChangedListener(listener.toNative())
     }
